Move domain stuff in the dal layer out of the dal layer.
